remove basicconfig
