+++ python/ray/signature.py
@@ -10,8 +10,8 @@ import logging
 from ray.utils import is_cython
 
 # Logger for this module. It should be configured at the entry point
-# into the program using Ray. Ray configures it by default automatically
-# using logging.basicConfig in its entry/init points.
+# into the program using Ray. Ray provides a default configuration at
+# entry/init points.
 logger = logging.getLogger(__name__)
 
 FunctionSignature = namedtuple("FunctionSignature", [
